AuthAPI App/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
passlib[bcrypt]
python-jose
pydantic
pyotp
qrcode
pytest
httpx
